avis/requirements.txt
grpcio
grpcio-tools
requests
pymongo
unidecode
ariadne
flask